Write ginkgo output to stderr
OTE was discarding the output ginkgo was writing to stdout -- this had
the side-effect of throwing away the test's output. We want that to be
stored in ExtensionTestResult.

Now that I'm running some longer running kube tests, it was also a
little weird we're silent for so long.  This changes the behavior so we
do include the tests output as it's generated, but it goes to stderr.

This should make it easier for origin to consume the JSON result from
stdout, but raw ginkgo test outputs still end up in the logs via stderr.

Note: this relies on a patch to ginkgo to expose configuring the writer,
stdout is hardcoded.